refactor: repository 인터페이스와 구현체로 구조 변경, api 주소 일괄 수정
